Add files via upload
Implement the set type which contains integers. Represent the set as a sequence of its
elements. Implement as methods: inserting an element, removing an element,
returning whether the set is empty, returning whether the set contains an element,
returning a random element without removing it from the set, returning the number of
even numbers in the set (suggestion: store the number of even numbers and update it
when the set changes), printing the set.